feat: Added "postPackPolygonScale" option for adding margins or overlap good
The (last parameter) options object for packAnimal now respects a postPackPolygonScale which
defaults to 1. It can be set to more than 1 to create overlap, or to less than 1 to create margins .
If you are using algorithmOptions.polygonHitboxScale you should probably switch to this as it is
much more reliable and has more predictable performance.

another #2 implementation