feat(staggerPack): introduce staggerPack
The introduction of staggerpack allows us to pack based on heuristics with fair certainty that we
get a pretty pack. Currently staggerPack will be used for sets of polygons with an aspect ratio
standard deviation below 0.25. For now there is no configurability and much room for improvement.
Also it looks pretty good.

BREAKING CHANGE: A large number of scenarios will now use staggerPack instead of greedyPack which
might present a big shift in your layouts. It's a move towards more aesthetic packing defaults